Add 1-array_iterator.c file

+++ function_pointers/1-array_iterator.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 
 /**
- * array_iterator - function that dynamically executes the function pointed by action
+ * array_iterator - dynamically executes the function pointed by action
  * @array: pointer to the array of integers
  * @size: size of the array
  * @action: pointer to the function to be executed
@@ -10,7 +10,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	long unsigned int i;
+	size_t i;
 
 	for (i = 0 ; i < size ; i++)
 		action(array[i]);
